usb: gadget: f_uac2: set UAC_EP_CS_ATTR_FILL_MAX in bmAttributes
For the capture endpoint, set the UAC_EP_CS_ATTR_FILL_MAX bit in
bmAttributes to that the host only send IN tokens with wMaxPacketSize
bytes. If that bit isn't set, snd-usb will calculate a smaller packet
size, which is answered by a fixed-size (512 bytes) response from the
gadget driver. That consequently leads to -EOVERFLOW errors on the
host side.

Reported-by: Sebastian Reimers <sebastian.reimers@googlemail.com>
Signed-off-by: Daniel Mack <zonque@gmail.com>