usb: gadget: f_uac2: tie capture clocking to system clock
The UAC2 function driver currently responds to all packets at all times
with wMaxPacketSize packets. That results in way too fast audio
playback as the function driver (which is in fact supposed to define
the audio stream pace) delivers as fast as it can.

Consider the following setup: the host uses arecord to record the
audio, and the gadget uses aplay to feed the stream. If both sides
use files as sink and source, everything is fine, because the data
is correctly transferred. However, in my tests, a one-minute wave
file took around 30 seconds to transfer that way. If one of the
sides expects data to be delivered at the nominal data rate though,
everything breaks apart.

Fix this by introducing some state variables to keep track of the
audio stream timing. Send 0-byte packets when the hosts ask for more
data than the local timing allows us to send.

Successfully tested on a BBB for the gadget side.

Signed-off-by: Daniel Mack <zonque@gmail.com>