bpf: add cgroup controller
This patch adds a cgroup controller that serves as container for two
sets of eBPF program pointers. One for such that are directly pinned
to a cgroup, and one for such that are effective for it.

To illustrate the logic behind that, assume the following example
cgroup hierarchy.

  A - B - C
        \ D - E

If only B has a program attached, it will be effective for B, C, D
and E. If D then attaches a program itself, that will be effective for
both D and E, and the program in B will only affect B and C. Only one
program of a given type is effective for a cgroup.

This controller does not expose any control knob files. Attaching and
detaching programs will be done through the bpf(2) syscall. For now,
ingress and egress inet socket filtering are the only supported
use-cases.

Signed-off-by: Daniel Mack <daniel@zonque.org>