power: supply: pda_power: add devicetree probing glue
This patch adds a devicetree probing table and some code to parse GPIOs that
may optionally be passed in through the bindings. The GPIOs are automatically
translated to IRQs.

The regulator name is changed to "ac" when probing is done via DT. The old
name is kept around for pdata compatibility.

Signed-off-by: Daniel Mack <daniel@zonque.org>